move: FeignClient의 DTO, 인터페이스 네이밍컨벤션에 맞게 수정
