+++ product-reservation-service/src/main/java/com/monkey/productreservationservice/infrastructure/feignclient/dto/response/ResStoreClientGetByIdDTOApiV1.java
@@ -1,0 +1,17 @@
+package com.monkey.productreservationservice.infrastructure.feignclient.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ResStoreClientGetByIdDTOApiV1 {
+    private UUID storeId;
+    private String storeName;
+}
