+++ product-service/src/main/java/com/monkey/productservice/application/dto/response/ResProductGetByIdDTOApiV1.java
@@ -1,8 +1,8 @@
 package com.monkey.productservice.application.dto.response;
 
 import com.monkey.productservice.domain.entity.ProductEntity;
-import com.monkey.productservice.infrastructure.feignclient.dto.StoreDTO;
-import com.monkey.productservice.infrastructure.feignclient.dto.UserDTO;
+import com.monkey.productservice.infrastructure.feignclient.dto.response.ResStoreClientGetByIdDTOApiV1;
+import com.monkey.productservice.infrastructure.feignclient.dto.response.ResUserClientGetByIdDTOApiV1;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,7 +17,7 @@ import java.util.UUID;
 public class ResProductGetByIdDTOApiV1 {
     private Product product;
 
-    public static ResProductGetByIdDTOApiV1 of(ProductEntity productEntity, StoreDTO storeDto, UserDTO userDto) {
+    public static ResProductGetByIdDTOApiV1 of(ProductEntity productEntity, ResStoreClientGetByIdDTOApiV1 storeDto, ResUserClientGetByIdDTOApiV1 userDto) {
         return ResProductGetByIdDTOApiV1.builder()
                 .product(Product.from(productEntity, storeDto, userDto))
                 .build();
@@ -36,7 +36,7 @@ public class ResProductGetByIdDTOApiV1 {
         private Store store;
         private User user;
 
-        public static Product from(ProductEntity productEntity, StoreDTO storeDto, UserDTO userDto) {
+        public static Product from(ProductEntity productEntity, ResStoreClientGetByIdDTOApiV1 storeDto, ResUserClientGetByIdDTOApiV1 userDto) {
             return Product.builder()
                     .productId(productEntity.getProductId())
                     .productName(productEntity.getProductName())
@@ -56,7 +56,7 @@ public class ResProductGetByIdDTOApiV1 {
             private UUID storeId;
             private String storeName;
 
-            public static Store from(StoreDTO storeDto) {
+            public static Store from(ResStoreClientGetByIdDTOApiV1 storeDto) {
                 return Store.builder()
                         .storeId(storeDto.getStoreId())
                         .storeName(storeDto.getStoreName())
@@ -71,7 +71,7 @@ public class ResProductGetByIdDTOApiV1 {
         public static class User {
             private Long userId;
 
-            public static User from(UserDTO userDto) {
+            public static User from(ResUserClientGetByIdDTOApiV1 userDto) {
                 return User.builder()
                         .userId(userDto.getUserId())
                         .build();
