+++ product-service/src/main/java/com/monkey/productservice/infrastructure/feignclient/StoreClientApiV1.java
@@ -1,7 +1,7 @@
 package com.monkey.productservice.infrastructure.feignclient;
 
 import com.monkey.common_module.dto.ResDTO;
-import com.monkey.productservice.infrastructure.feignclient.dto.StoreDTO;
+import com.monkey.productservice.infrastructure.feignclient.dto.response.ResStoreClientGetByIdDTOApiV1;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import java.util.UUID;
 
 @FeignClient(name = "store-service", path = "v1/stores")
-public interface StoreFeignClient {
+public interface StoreClientApiV1 {
     @GetMapping
-    ResDTO<StoreDTO> getStoreById(@PathVariable("storeId") UUID storeId);
+    ResDTO<ResStoreClientGetByIdDTOApiV1> getStoreById(@PathVariable("storeId") UUID storeId);
 }
