--- product-service/src/main/java/com/monkey/productservice/infrastructure/feignclient/dto/UserDTO.java
@@ -1,14 +1,0 @@
-package com.monkey.productservice.infrastructure.feignclient.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserDTO {
-    private Long userId;
-}
