+++ product-service/src/main/java/com/monkey/productservice/infrastructure/feignclient/dto/response/ResStoreClientGetByIdDTOApiV1.java
@@ -1,4 +1,4 @@
-package com.monkey.productreservationservice.infrastructure.feignclient.dto;
+package com.monkey.productservice.infrastructure.feignclient.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
