+++ product-service/src/main/java/com/monkey/productservice/infrastructure/feignclient/dto/response/ResUserClientGetByIdDTOApiV1.java
@@ -1,17 +1,14 @@
-package com.monkey.productservice.infrastructure.feignclient.dto;
+package com.monkey.productservice.infrastructure.feignclient.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.UUID;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class StoreDTO {
-    private UUID storeId;
-    private String storeName;
+public class ResUserClientGetByIdDTOApiV1 {
+    private Long userId;
 }
